+++ src/main/java/ru/milov/transactions/service/domain/UserBill.java
@@ -2,6 +2,7 @@ package ru.milov.transactions.service.domain;
 
 import lombok.Getter;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -10,12 +11,10 @@ public class UserBill {
     private Integer user_id;
     private Integer bill_id;
     private String name;
-    private int balance;
+    private BigDecimal balance;
 
     @Override
     public String toString() {
         return name + '\t' + balance ;
-
-
     }
 }
